pyproject.toml
[tool.poetry]
name = "deepsushi"
version = "0.1.0"
description = ""
authors = ["D. Adamson <0xwebmilk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-decouple = "^3.8"
pydantic = "^2.9.2"
flask = "^3.0.3"
flask-jwt-extended = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
